requirements.txt
dj-database-url==0.5.0
Django==4.2.7
django-heroku==0.3.1
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
gunicorn==21.2.0
psycopg2==2.9.9
pytz==2018.5
shortuuid==0.5.0
six==1.11.0
whitenoise==6.6.0
django-cors-headers==4.3.1
requests==2.31.0
